+++ src/routes/graphql/index.ts
@@ -1,24 +1,20 @@
 import { FastifyPluginAsyncJsonSchemaToTs } from "@fastify/type-provider-json-schema-to-ts";
 import { graphql, parse, validate } from "graphql";
 import depthLimit = require("graphql-depth-limit");
-import {
-  GraphQLInt,
-  GraphQLObjectType,
-  GraphQLFloat,
-  GraphQLString,
-  GraphQLSchema,
-  GraphQLList,
-  GraphQLInputObjectType,
-  GraphQLNonNull,
-} from "graphql/type";
-import { MemberTypeEntity } from "../../utils/DB/entities/DBMemberTypes";
-import { PostEntity } from "../../utils/DB/entities/DBPosts";
-import { ProfileEntity } from "../../utils/DB/entities/DBProfiles";
-import { UserEntity } from "../../utils/DB/entities/DBUsers";
+import { GraphQLID, GraphQLSchema } from "graphql/type";
 import { getLoaders } from "./loaders";
+import { Mutations } from "./mutation";
+import { RootQuery } from "./query";
 import { graphqlBodySchema } from "./schema";
 import { Context } from "./types/context";
-import { GraphQLUUID } from "./types/uuid";
+import { MemberType, UpdateMemberType } from "./types/memberType";
+import { PostType, CreatePostType, UpdatePostType } from "./types/post";
+import {
+  ProfileType,
+  CreateProfileType,
+  UpdateProfileType,
+} from "./types/profile";
+import { UserType, CreateUserType, UpdateUserType } from "./types/user";
 
 const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
   fastify
@@ -30,8 +26,8 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
         body: graphqlBodySchema,
       },
     },
-    async function (request, reply) {
-      const { query, mutation, variables } = request.body;
+    async function (request) {
+      const { query, mutation } = request.body;
       const source: string | null = query || mutation || null;
 
       if (!source) throw this.httpErrors.badRequest();
@@ -48,7 +44,6 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       return graphql({
         schema,
         source,
-        variableValues: variables,
         contextValue,
       });
     }
@@ -57,351 +52,21 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
 
 export default plugin;
 
-const memberType: GraphQLObjectType = new GraphQLObjectType({
-  name: "MemberType",
-  fields: () => ({
-    id: { type: GraphQLString },
-    discount: { type: new GraphQLNonNull(GraphQLFloat) },
-    monthPostsLimit: { type: new GraphQLNonNull(GraphQLInt) },
-  }),
-});
-
-const updateMemberType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "UpdateMemberTypeDTO",
-  fields: () => ({
-    discount: { type: GraphQLFloat },
-    monthPostsLimit: { type: GraphQLInt },
-  }),
-});
-
-const postType: GraphQLObjectType = new GraphQLObjectType({
-  name: "Post",
-  fields: () => ({
-    id: { type: GraphQLUUID },
-    title: { type: new GraphQLNonNull(GraphQLString) },
-    content: { type: new GraphQLNonNull(GraphQLString) },
-    author: {
-      type: new GraphQLNonNull(userType),
-      resolve: (post: PostEntity, args, ctx: Context) => {
-        return ctx.userLoader.load(post.userId);
-      },
-    },
-  }),
-});
-
-const createPostType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "CreatePostDTO",
-  fields: () => ({
-    title: { type: new GraphQLNonNull(GraphQLString) },
-    content: { type: new GraphQLNonNull(GraphQLString) },
-    userId: { type: new GraphQLNonNull(GraphQLUUID) },
-  }),
-});
-
-const updatePostType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "UpdatePostDTO",
-  fields: () => ({
-    title: { type: GraphQLString },
-    content: { type: GraphQLString },
-  }),
-});
-
-const userType: GraphQLObjectType = new GraphQLObjectType({
-  name: "User",
-  fields: () => ({
-    id: { type: GraphQLUUID },
-    firstName: { type: new GraphQLNonNull(GraphQLString) },
-    lastName: { type: new GraphQLNonNull(GraphQLString) },
-    email: { type: new GraphQLNonNull(GraphQLString) },
-    profile: {
-      type: profileType,
-      resolve: (user: UserEntity, args, ctx: Context) => {
-        return ctx.profileByUserIdLoader.load(user.id);
-      },
-    },
-    posts: {
-      type: new GraphQLList(postType),
-      resolve: (user: UserEntity, args, ctx: Context) => {
-        return ctx.fastify.db.posts.findMany({
-          key: "userId",
-          equals: user.id,
-        });
-      },
-    },
-    userSubscribedTo: {
-      type: new GraphQLList(userType),
-      resolve: (user: UserEntity, args, ctx: Context) => {
-        return ctx.fastify.db.users.findMany({
-          key: "subscribedToUserIds",
-          inArray: user.id,
-        });
-      },
-    },
-    subscribedToUser: {
-      type: new GraphQLList(userType),
-      resolve: (user: UserEntity, args, ctx: Context) => {
-        return ctx.fastify.db.users.findMany({
-          key: "id",
-          equalsAnyOf: user.subscribedToUserIds,
-        });
-      },
-    },
-  }),
-});
-
-const createUserType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "CreateUserDTO",
-  fields: () => ({
-    firstName: { type: new GraphQLNonNull(GraphQLString) },
-    lastName: { type: new GraphQLNonNull(GraphQLString) },
-    email: { type: new GraphQLNonNull(GraphQLString) },
-  }),
-});
-
-const updateUserType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "UpdateUserDTO",
-  fields: () => ({
-    firstName: { type: GraphQLString },
-    lastName: { type: GraphQLString },
-    email: { type: GraphQLString },
-  }),
-});
-
-const profileType: GraphQLObjectType = new GraphQLObjectType({
-  name: "Profile",
-  fields: () => ({
-    id: { type: GraphQLUUID },
-    avatar: { type: new GraphQLNonNull(GraphQLString) },
-    sex: { type: new GraphQLNonNull(GraphQLString) },
-    birthday: { type: new GraphQLNonNull(GraphQLInt) },
-    country: { type: new GraphQLNonNull(GraphQLString) },
-    street: { type: new GraphQLNonNull(GraphQLString) },
-    city: { type: new GraphQLNonNull(GraphQLString) },
-    user: { type: new GraphQLNonNull(userType) },
-    memberType: {
-      type: new GraphQLNonNull(memberType),
-      resolve: async (profile: ProfileEntity, args, ctx: Context) => {
-        return ctx.memberTypeLoader.load(profile.memberTypeId);
-      },
-    },
-  }),
-});
-
-const createProfileType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "CreateProfileDTO",
-  fields: () => ({
-    avatar: { type: new GraphQLNonNull(GraphQLString) },
-    sex: { type: new GraphQLNonNull(GraphQLString) },
-    birthday: { type: new GraphQLNonNull(GraphQLInt) },
-    country: { type: new GraphQLNonNull(GraphQLString) },
-    street: { type: new GraphQLNonNull(GraphQLString) },
-    city: { type: new GraphQLNonNull(GraphQLString) },
-    userId: { type: new GraphQLNonNull(GraphQLUUID) },
-    memberTypeId: { type: new GraphQLNonNull(GraphQLString) },
-  }),
-});
-
-const updateProfileType: GraphQLInputObjectType = new GraphQLInputObjectType({
-  name: "UpdateProfileDTO",
-  fields: () => ({
-    avatar: { type: GraphQLString },
-    sex: { type: GraphQLString },
-    birthday: { type: GraphQLInt },
-    country: { type: GraphQLString },
-    street: { type: GraphQLString },
-    city: { type: GraphQLString },
-    userId: { type: GraphQLUUID },
-    memberTypeId: { type: GraphQLString },
-  }),
-});
-
-export const RootQuery = new GraphQLObjectType({
-  name: "RootQueryType",
-  fields: {
-    memberTypes: {
-      type: new GraphQLList(memberType),
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.fastify.db.memberTypes.findMany();
-      },
-    },
-    memberType: {
-      type: memberType,
-      args: {
-        id: { type: GraphQLString },
-      },
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.memberTypeLoader.load(args.id);
-      },
-    },
-    users: {
-      type: new GraphQLList(userType),
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.fastify.db.users.findMany();
-      },
-    },
-    user: {
-      type: userType,
-      args: {
-        id: { type: GraphQLUUID },
-      },
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.userLoader.load(args.id);
-      },
-    },
-    posts: {
-      type: new GraphQLList(postType),
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.fastify.db.posts.findMany();
-      },
-    },
-    post: {
-      type: postType,
-      args: {
-        id: { type: GraphQLUUID },
-      },
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.postLoader.load(args.id);
-      },
-    },
-    profiles: {
-      type: new GraphQLList(profileType),
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.fastify.db.profiles.findMany();
-      },
-    },
-    profile: {
-      type: profileType,
-      args: {
-        id: { type: GraphQLUUID },
-      },
-      resolve: (obj, args, ctx: Context) => {
-        return ctx.profileLoader.load(args.id);
-      },
-    },
-  },
-});
-
-const Mutations = new GraphQLObjectType({
-  name: "Mutations",
-  fields: {
-    createUser: {
-      type: userType,
-      args: {
-        createUserDTO: { type: createUserType },
-      },
-      async resolve(obj, args, ctx: Context) {
-        return ctx.fastify.db.users.create(args.createUserDTO);
-      },
-    },
-    createProfile: {
-      type: profileType,
-      args: {
-        createProfileDTO: { type: createProfileType },
-      },
-      async resolve(obj, args, ctx: Context) {
-        return ctx.fastify.db.profiles.create(args.createProfileDTO);
-      },
-    },
-    createPost: {
-      type: postType,
-      args: {
-        createPostDTO: { type: createPostType },
-      },
-      async resolve(obj, args, ctx: Context) {
-        return ctx.fastify.db.posts.create(args.createPostDTO);
-      },
-    },
-    updateMemberType: {
-      type: memberType,
-      args: {
-        id: { type: new GraphQLNonNull(GraphQLString) },
-        updateMemberTypeDTO: { type: updateMemberType },
-      },
-      async resolve(memberType: MemberTypeEntity, args, ctx: Context) {
-        return ctx.fastify.db.memberTypes.change(
-          args.id,
-          args.updateMemberTypeDTO
-        );
-      },
-    },
-    updatePost: {
-      type: postType,
-      args: {
-        id: { type: new GraphQLNonNull(GraphQLUUID) },
-        updatePostDTO: { type: updatePostType },
-      },
-      async resolve(post: PostEntity, args, ctx: Context) {
-        return ctx.fastify.db.posts.change(args.id, args.updatePostDTO);
-      },
-    },
-    updateProfile: {
-      type: profileType,
-      args: {
-        id: { type: new GraphQLNonNull(GraphQLUUID) },
-        updateProfileDTO: { type: updateProfileType },
-      },
-      async resolve(_, args, ctx: Context) {
-        return ctx.fastify.db.profiles.change(args.id, args.updateProfileDTO);
-      },
-    },
-    updateUser: {
-      type: userType,
-      args: {
-        id: { type: new GraphQLNonNull(GraphQLUUID) },
-        updateUserDTO: { type: updateUserType },
-      },
-      async resolve(_, args, ctx: Context) {
-        return ctx.fastify.db.users.change(args.id, args.updateUserDTO);
-      },
-    },
-    subscribeTo: {
-      type: userType,
-      args: {
-        userId: { type: new GraphQLNonNull(GraphQLUUID) },
-        subscriberId: { type: new GraphQLNonNull(GraphQLUUID) },
-      },
-      resolve: async (_, args, ctx: Context) => {
-        const { userId, subscriberId } = args;
-        const user = await ctx.userLoader.load(userId);
-        const subscriber = await ctx.userLoader.load(subscriberId);
-        if (
-          user &&
-          subscriber &&
-          !user.subscribedToUserIds.includes(subscriberId)
-        ) {
-          const subscribedToUserIds = [
-            ...user.subscribedToUserIds,
-            subscriberId,
-          ];
-          return await ctx.fastify.db.users.change(userId, {
-            subscribedToUserIds,
-          });
-        }
-      },
-    },
-    unsubscribeFrom: {
-      type: userType,
-      args: {
-        userId: { type: new GraphQLNonNull(GraphQLUUID) },
-        unsubId: { type: new GraphQLNonNull(GraphQLUUID) },
-      },
-      resolve: async (_, args, ctx: Context) => {
-        const { userId, unsubId } = args;
-        const user = await ctx.userLoader.load(userId);
-        const unsub = await ctx.userLoader.load(unsubId);
-
-        if (user && unsub && user.subscribedToUserIds.includes(unsubId)) {
-          const subscribedToUserIds = user.subscribedToUserIds.filter(
-            (id) => id !== unsubId
-          );
-          return ctx.fastify.db.users.change(userId, { subscribedToUserIds });
-        }
-      },
-    },
-  },
-});
-
 export const schema = new GraphQLSchema({
   query: RootQuery,
   mutation: Mutations,
+  types: [
+    UserType,
+    ProfileType,
+    PostType,
+    MemberType,
+    CreateUserType,
+    CreatePostType,
+    CreateProfileType,
+    UpdateMemberType,
+    UpdateUserType,
+    UpdatePostType,
+    UpdateProfileType,
+    GraphQLID,
+  ],
 });
